requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.7.22
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
crispy-bootstrap5==0.7
cycler==0.12.1
Django==4.2.4
django-appconf==1.0.5
django-autocomplete-light==3.9.7
django-cors-headers==4.3.0
django-crispy-forms==2.1
django-filter==23.3
django-jet-reboot==1.3.3
django-leaflet==0.29.0
django-pandas==0.6.6
djangorestframework==3.14.0
docopt==0.6.2
Flask==3.0.0
fonttools==4.44.0
frozenlist==1.4.0
GDAL @ file:///C:/Users/sougrinoma/pipwin/GDAL-3.4.3-cp311-cp311-win_amd64.whl#sha256=f78861fb5115d5c2f8cf3c52a492ff548da9e1256dc84088947379f90e77e5b6
geoip2==4.7.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Js2Py==0.74
kiwisolver==1.4.5
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.1
maxminddb==2.5.1
multidict==6.0.4
num2words==0.5.13
numpy==1.26.1
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
pipwin==0.5.2
plotly==5.18.0
psycopg2-binary==2.9.9
pyjsparser==2.7.1
PyJWT==2.8.0
pyparsing==3.1.1
PyPrind==2.11.3
pyproj==3.6.1
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
schedule==1.2.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tenacity==8.2.3
twilio==8.10.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.4
Werkzeug==3.0.1
yarl==1.9.2
zipp==3.17.0
